refactor(providers,virtualmodels,mcpgateway): serve both SQL backends from one store
Three config-row stores with the same List/Get/Upsert/Delete shape.

mcpgateway's display_name migration read PRAGMA table_info to decide whether
to add the column, which only works on SQLite; it becomes an AddColumns entry
plus the backfill UPDATE, and the migration test now runs on both dialects
from the pre-display_name table shape.

virtualmodels' UpsertAll atomicity test needed two independent databases in
one run, so it splits into a commit case and a rollback case — each gets its
own database from sqlxtest, which reads better than the two-stores-in-one-test
form it replaces.
