+++ internal/virtualmodels/factory.go
@@ -2,17 +2,16 @@ package virtualmodels
 
 import (
 	"context"
-	"database/sql"
 	"errors"
 	"fmt"
 	"sync"
 	"time"
 
-	"github.com/jackc/pgx/v5/pgxpool"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 
 	"github.com/enterpilot/gomodel/config"
 	"github.com/enterpilot/gomodel/internal/storage"
+	"github.com/enterpilot/gomodel/internal/storage/sqlx"
 )
 
 // Result holds the initialized virtual models service and any owned resources.
@@ -131,10 +130,10 @@ func newResult(ctx context.Context, cfg *config.Config, storeConn storage.Storag
 }
 
 func createStore(ctx context.Context, store storage.Storage) (Store, error) {
-	return storage.ResolveBackend[Store](
+	return storage.ResolveSQLBackend[Store](
+		ctx,
 		store,
-		func(db *sql.DB) (Store, error) { return NewSQLiteStore(db) },
-		func(pool *pgxpool.Pool) (Store, error) { return NewPostgreSQLStore(ctx, pool) },
+		func(db sqlx.DB) (Store, error) { return NewSQLStore(ctx, db) },
 		func(db *mongo.Database) (Store, error) { return NewMongoDBStore(db) },
 	)
 }
